chore : Album에 관련된 extention 생성
